-introduced app component as a root for the docs -reverted commenting of the merged variables in the variables panel
